Create gui.java
Design and implement a GUI for the Temperature class. One challenge of this design is to find a good way for the user to indicate whether a Fahrenheit or Celsius value is being input. This should also determine the order of the conversion: F to C or C to F. [C/5 = (F -32)/9]